Reorg with partial classes
