+++ Emitter/Emitter.cs
@@ -50,11 +50,12 @@ namespace Emitter
     /// <summary>
     /// Represents emitter.io MQTT-based client.
     /// </summary>
-    public class Connection : IDisposable
+    public partial class Connection : IDisposable
     {
         #region Constructors
         private readonly MqttClient Client;
         private readonly ReverseTrie Trie = new ReverseTrie(-1);
+        private readonly ReverseTrie PresenceTrie = new ReverseTrie(-1);
         private string DefaultKey = null;
 
         /// <summary>
@@ -210,257 +211,8 @@ namespace Emitter
 
         #endregion Connect / Disconnect Members
 
-        #region Subscribe / Unsubscribe Members
+        //public event PresenceHandler Presence;
 
-        /// <summary>
-        /// Asynchronously subscribes to a particular channel of emitter.io service. Uses the default
-        /// key that should be specified in the constructor.
-        /// </summary>
-        /// <param name="channel">The channel to subscribe to.</param>
-        /// <param name="handler">The callback to be invoked every time the message is received.</param>
-        /// <returns>The message identifier for this operation.</returns>
-        public ushort On(string channel, MessageHandler handler)
-        {
-            if (this.DefaultKey == null)
-                throw EmitterException.NoDefaultKey;
-            return this.On(this.DefaultKey, channel, handler);
-        }
-
-        /// <summary>
-        /// Asynchronously subscribes to a particular channel of emitter.io service.
-        /// </summary>
-        /// <param name="key">The key to use for this subscription request.</param>
-        /// <param name="channel">The channel to subscribe to.</param>
-        /// <param name="handler">The callback to be invoked every time the message is received.</param>
-        /// <returns>The message identifier for this operation.</returns>
-        public ushort On(string key, string channel, MessageHandler handler)
-        {
-            // Register the handler
-            this.Trie.RegisterHandler(channel, handler);
-
-            // Subscribe
-            return this.Client.Subscribe(new string[] { FormatChannel(key, channel) }, new byte[] { MqttMsgBase.QOS_LEVEL_AT_MOST_ONCE });
-        }
-
-        /// <summary>
-        /// Asynchronously subscribes to a particular channel of emitter.io service.
-        /// </summary>
-        /// <param name="key">The key to use for this subscription request.</param>
-        /// <param name="channel">The channel to subscribe to.</param>
-        /// <param name="handler">The callback to be invoked every time the message is received.</param>
-        /// <param name="last">The last x messages to request when we subcribe.</param>
-        /// <returns>The message identifier for this operation.</returns>
-        public ushort On(string key, string channel, MessageHandler handler, int last)
-        {
-            // Register the handler
-            this.Trie.RegisterHandler(channel, handler);
-
-            // Subscribe
-            return this.Client.Subscribe(new string[] { FormatChannel(key, channel, Options.WithLast(last)) }, new byte[] { MqttMsgBase.QOS_LEVEL_AT_MOST_ONCE });
-        }
-
-        /// <summary>
-        /// Asynchonously unsubscribes from a particular channel of emitter.io service. Uses the default
-        /// key that should be specified in the constructor.
-        /// </summary>
-        /// <param name="channel">The channel to subscribe to.</param>
-        /// <returns>The message identifier for this operation.</returns>
-        public ushort Unsubscribe(string channel)
-        {
-            if (this.DefaultKey == null)
-                throw EmitterException.NoDefaultKey;
-            return this.Unsubscribe(this.DefaultKey, channel);
-        }
-
-        /// <summary>
-        /// Asynchonously unsubscribes from a particular channel of emitter.io service.
-        /// </summary>
-        /// <param name="key">The key to use for this unsubscription request.</param>
-        /// <param name="channel">The channel to subscribe to.</param>
-        /// <returns>The message identifier for this operation.</returns>
-        public ushort Unsubscribe(string key, string channel)
-        {
-            // Unregister the handler
-            this.Trie.UnregisterHandler(key);
-
-            // Unsubscribe
-            return this.Client.Unsubscribe(new string[] { FormatChannel(key, channel) });
-        }
-
-        #endregion Subscribe / Unsubscribe Members
-
-        #region Publish Members
-
-        /// <summary>
-        /// Asynchonously publishes a message to the emitter.io service. Uses the default
-        /// key that should be specified in the constructor.
-        /// </summary>
-        /// <param name="channel">The channel to publish to.</param>
-        /// <param name="message">The message body to send.</param>
-        /// <returns>The message identifier for this operation.</returns>
-        public ushort Publish(string channel, byte[] message)
-        {
-            if (this.DefaultKey == null)
-                throw EmitterException.NoDefaultKey;
-            return this.Publish(this.DefaultKey, channel, message);
-        }
-
-        /// <summary>
-        /// Asynchonously publishes a message to the emitter.io service. Uses the default
-        /// key that should be specified in the constructor.
-        /// </summary>
-        /// <param name="channel">The channel to publish to.</param>
-        /// <param name="message">The message body to send.</param>
-        /// <returns>The message identifier for this operation.</returns>
-        public ushort Publish(string channel, string message)
-        {
-            if (this.DefaultKey == null)
-                throw EmitterException.NoDefaultKey;
-            return this.Publish(this.DefaultKey, channel, Encoding.UTF8.GetBytes(message));
-        }
-
-        /// <summary>
-        /// Publishes a message to the emitter.io service asynchronously.
-        /// </summary>
-        /// <param name="key">The key to use for this publish request.</param>
-        /// <param name="channel">The channel to publish to.</param>
-        /// <param name="message">The message body to send.</param>
-        /// <returns>The message identifier.</returns>
-        public ushort Publish(string key, string channel, string message)
-        {
-            return this.Client.Publish(FormatChannel(key, channel), Encoding.UTF8.GetBytes(message));
-        }
-
-        /// <summary>
-        /// Publishes a message to the emitter.io service asynchronously.
-        /// </summary>
-        /// <param name="key">The key to use for this publish request.</param>
-        /// <param name="channel">The channel to publish to.</param>
-        /// <param name="message">The message body to send.</param>
-        /// <returns>The message identifier.</returns>
-        public ushort Publish(string key, string channel, byte[] message)
-        {
-            return this.Client.Publish(FormatChannel(key, channel), message);
-        }
-
-        /// <summary>
-        /// Publishes a message to the emitter.io service asynchronously.
-        /// </summary>
-        /// <param name="key">The key to use for this publish request.</param>
-        /// <param name="channel">The channel to publish to.</param>
-        /// <param name="message">The message body to send.</param>
-        /// <param name="ttl">The time to live of the message.</param>
-        /// <returns>The message identifier.</returns>
-        public ushort Publish(string key, string channel, string message, int ttl)
-        {
-            return this.Client.Publish(FormatChannel(key, channel, Options.WithTTL(ttl)), Encoding.UTF8.GetBytes(message));
-        }
-
-        /// <summary>
-        /// Publishes a message to the emitter.io service asynchronously.
-        /// </summary>
-        /// <param name="key">The key to use for this publish request.</param>
-        /// <param name="channel">The channel to publish to.</param>
-        /// <param name="message">The message body to send.</param>
-        /// <param name="options">The options associated with the message, such as 'ttl' or 'last'. Ex: Options.WithLast(5).</param>
-        /// <returns>The message identifier.</returns>
-        public ushort Publish(string key, string channel, string message, params string[] options)
-        {
-            GetHeader(options, out var retain, out var qos);
-            return this.Client.Publish(FormatChannel(key, channel, options), Encoding.UTF8.GetBytes(message), qos, retain);
-        }
-
-        /// <summary>
-        /// Publishes a message to the emitter.io service asynchronously.
-        /// </summary>
-        /// <param name="key">The key to use for this publish request.</param>
-        /// <param name="channel">The channel to publish to.</param>
-        /// <param name="message">The message body to send.</param>
-        /// <param name="ttl">The time to live of the message.</param>
-        /// <returns>The message identifier.</returns>
-        public ushort Publish(string key, string channel, byte[] message, int ttl)
-        {
-            return this.Client.Publish(FormatChannel(key, channel, Options.WithTTL(ttl)), message);
-        }
-
-        /// <summary>
-        /// Publishes a message to the emitter.io service asynchronously.
-        /// </summary>
-        /// <param name="key">The key to use for this publish request.</param>
-        /// <param name="channel">The channel to publish to.</param>
-        /// <param name="message">The message body to send.</param>
-        /// <param name="options">The options associated with the message, such as 'ttl' or 'last'. Ex: Options.WithLast(5).</param>
-        /// <returns>The message identifier.</returns>
-        public ushort Publish(string key, string channel, byte[] message, params string[] options)
-        {
-            GetHeader(options, out var retain, out var qos);
-            return this.Client.Publish(FormatChannel(key, channel, options), message, qos, retain);
-        }
-
-        #endregion Publish Members
-
-        #region KeyGen Members
-
-        /// <summary>
-        /// Hashtable used for processing keygen responses.
-        /// </summary>
-        private readonly Hashtable KeygenHandlers = new Hashtable();
-
-        /// <summary>
-        /// Asynchronously sends a key generation request to the emitter.io service.
-        /// </summary>
-        /// <param name="secretKey">The secret key for this request.</param>
-        /// <param name="channel">The target channel for the requested key.</param>
-        /// <param name="securityAccess">The security access of the requested key.</param>
-        public void GenerateKey(string secretKey, string channel, SecurityAccess securityAccess, KeygenHandler handler)
-        {
-            this.GenerateKey(secretKey, channel, securityAccess, 0, handler);
-        }
-
-        /// <summary>
-        /// Asynchronously sends a key generation request to the emitter.io service.
-        /// </summary>
-        /// <param name="secretKey">The secret key for this request.</param>
-        /// <param name="channel">The target channel for the requested key.</param>
-        /// <param name="securityAccess">The security access of the requested key.</param>
-        /// <param name="ttl">The number of seconds for which this key will be usable.</param>
-        public void GenerateKey(string secretKey, string channel, SecurityAccess securityAccess, int ttl, KeygenHandler handler)
-        {
-            // Prepare the request
-            var request = new KeygenRequest();
-            request.Key = secretKey;
-            request.Channel = channel;
-            request.Type = securityAccess;
-            request.Ttl = ttl;
-
-            // Register the handler
-            this.KeygenHandlers[channel] = handler;
-
-            //this.Client.Subscribe(new string[] { "emitter/keygen/" }, new byte[] { 0 });
-
-            // Serialize and publish the request
-            this.Publish("emitter/", "keygen/", Encoding.UTF8.GetBytes(request.ToJson()));
-        }
-
-        #endregion KeyGen Members
-
-        #region Presence Members
-
-        public event PresenceHandler Presence;
-
-        public void PresenceSubscription(string key, string channel, bool status, bool changes)
-        {
-            var request = new PresenceRequest();
-            request.Key = key;
-            request.Channel = channel;
-            request.Status = status;
-            request.Changes = changes;
-
-            this.Publish("emitter/", "presence", Encoding.UTF8.GetBytes(request.ToJson()));
-        }
-        #endregion Presence Members
-
-        #region Private Members
 
         /// <summary>
         /// Occurs when a message is received.
@@ -509,7 +261,7 @@ namespace Emitter
                 this.InvokeError(ex);
             }
         }
-
+        #region Private Members
         /// <summary>
         /// Formats the channel.
         /// </summary>
@@ -601,158 +353,4 @@ namespace Emitter
 
         #endregion IDisposable
     }
-
-    #region ReverseTrie
-
-    /// <summary>
-    /// Represents a trie with a reverse-pattern search.
-    /// </summary>
-    /// <typeparam name="T"></typeparam>
-    internal class ReverseTrie
-    {
-        private readonly Hashtable Children;
-        private readonly short Level = 0;
-        private MessageHandler Value = default(MessageHandler);
-
-        /// <summary>
-        /// Constructs a node of the trie.
-        /// </summary>
-        /// <param name="level">The level of this node within the trie.</param>
-        public ReverseTrie(short level)
-        {
-            this.Level = level;
-            this.Children = new Hashtable();
-        }
-
-        /// <summary>
-        /// Adds a new handler for the channel.
-        /// </summary>
-        /// <param name="channel"></param>
-        /// <param name="value"></param>
-        public void RegisterHandler(string channel, MessageHandler value)
-        {
-            // Add the value or replace it.
-            this.AddOrUpdate(CreateKey(channel), 0, () => value, (old) => value);
-        }
-
-        /// <summary>
-        /// Unregister the handler from the trie.
-        /// </summary>
-        /// <param name="channel"></param>
-        public void UnregisterHandler(string channel)
-        {
-            MessageHandler removed;
-            this.TryRemove(CreateKey(channel), 0, out removed);
-        }
-
-        /// <summary>
-        /// Retrieves a set of values.
-        /// </summary>
-        /// <param name="query">The query to retrieve.</param>
-        /// <param name="position">The position.</param>
-        /// <returns></returns>
-        public IEnumerable Match(string channel)
-        {
-            // Matches
-            var result = new ArrayList();
-
-            // Get the query
-            var query = CreateKey(channel);
-
-            // Get the matching stack
-            var matches = new Stack();
-
-            // Push the root
-            object childNode;
-            matches.Push(this);
-            while (matches.Count != 0)
-            {
-                var current = matches.Pop() as ReverseTrie;
-                if (current.Value != default(object))
-                    result.Add(current.Value);
-
-                var level = current.Level + 1;
-                if (level >= query.Length)
-                    break;
-
-                if (Utils.TryGetValueFromHashtable(current.Children, "+", out childNode))
-                    matches.Push(childNode);
-                if (Utils.TryGetValueFromHashtable(current.Children, query[level], out childNode))
-                    matches.Push(childNode);
-            }
-
-            return result;
-        }
-
-        #region Private Members
-
-        /// <summary>
-        /// Creates a query for the trie from the channel name.
-        /// </summary>
-        /// <param name="channel"></param>
-        /// <returns></returns>
-        public static string[] CreateKey(string channel)
-        {
-            return channel.Split('/');
-        }
-
-        /// <summary>
-        /// Adds or updates a specific value.
-        /// </summary>
-        /// <param name="key"></param>
-        /// <param name="position"></param>
-        /// <returns></returns>
-        private object AddOrUpdate(string[] key, int position, AddFunc addFunc, UpdateFunc updateFunc)
-        {
-            if (position >= key.Length)
-            {
-                lock (this)
-                {
-                    // There's already a value
-                    if (this.Value != default(object))
-                        return updateFunc(this.Value);
-
-                    // No value, add it
-                    this.Value = addFunc();
-                    return this.Value;
-                }
-            }
-
-            // Create a child
-            var child = Utils.GetOrAddToHashtable(Children, key[position], new ReverseTrie((short)position)) as ReverseTrie;
-            return child.AddOrUpdate(key, position + 1, addFunc, updateFunc);
-        }
-
-        /// <summary>
-        /// Attempts to remove a specific key from the Trie.
-        /// </summary>
-        private bool TryRemove(string[] key, int position, out MessageHandler value)
-        {
-            if (position >= key.Length)
-            {
-                lock (this)
-                {
-                    // There's no value
-                    value = this.Value;
-                    if (this.Value == default(MessageHandler))
-                        return false;
-
-                    this.Value = default(MessageHandler);
-                    return true;
-                }
-            }
-
-            // Remove from the child
-            object child;
-            if (Utils.TryGetValueFromHashtable(Children, key[position], out child))
-                return ((ReverseTrie)child).TryRemove(key, position + 1, out value);
-
-            value = default(MessageHandler);
-            return false;
-        }
-
-        #endregion Private Members
-    }
-
-    #endregion ReverseTrie
 }
